fix(emit): resolve interior let-bound recur values in the int-safety check
Follow-up to 5facc65 (the loop-variable I32 boxing bug fix): array.prn's
own matmul/transpose were still producing wrong numbers even after
that fix landed. Root-caused to a second, separate bug -- this time in
the fix's own safety-net check, not the original bug: strides-for's
own accumulator (`running`/`next`) is reassigned via a recur value that
is a bare reference to an interior `let`-bound local (`next`), and
loop_recur_binding_is_int_safe's own `let` case recursed into the
let's body using the SAME outer scope, never actually binding `next`
into it -- so the check's own emit_expr dry-run always failed to
resolve `next`'s type, conservatively (and incorrectly) downgrading
the whole chain back to `double`.

Fixed by giving that `let` case a real, properly-scoped child
EmitScope, mirroring emit_let's own real binding loop (resolve each
init's type via the same emit_expr, scope_bind it) before recursing
into the let's own body -- a pure, scope-populating dry run, no C ever
kept or emitted.

Verified: matmul(A,B) and transpose(A) now produce the real,
hand-computed correct values end to end (a genuine 2x3 * 3x2 product,
[[58,64],[139,154]]) -- previously long-documented in STDLIB.md/
linalg.prn as "compiles clean but NOT verified numerically correct."
New make test-linalg-matmul, the real, permanent regression gate.

make test: 347/347. Full real-target regression sweep re-run clean,
zero regressions (same set as commit 5facc65's own sweep).

backlog: ✓ PARENA-ARRAY-MATMUL-001


session: sess-20260905-0720-ec33e7c5

+++ tests/test_ami_gen.c
@@ -254,7 +254,7 @@ Result parse_message(char * raw __attribute__((unused)), Arena *dest __attribute
     int n __attribute__((unused)) = length(raw);
     Vec headers __attribute__((unused)) = vec_new(dest);
     Result __loop_result_4 __attribute__((unused));
-    double i = 0;
+    int i = 0;
     while (1) {
         int line_end __attribute__((unused)) = find_crlf(raw, i, n);
         if ((i >= n)) {
@@ -273,7 +273,7 @@ Result parse_message(char * raw __attribute__((unused)), Arena *dest __attribute
         char *hname __attribute__((unused)) = substring(raw, i, colon, dest);
         char *hval __attribute__((unused)) = substring(raw, (colon + 2), line_end, dest);
         (void)(vec_push_(&(headers), AmiHeader_box(dest, AmiHeader_new(hname, hval))));
-        double __recur_tmp_0 = (line_end + 2);
+        int __recur_tmp_0 = (line_end + 2);
         i = __recur_tmp_0;
         continue;
         }
